Re-structured the project
Restructured the project to make sense, removed redundant code
