+++ README.md
@@ -1,0 +1,7 @@
+# Neural-Network-
+
+## How to execution
+
+- javac *.java
+
+- java NeuralNetwork
