--- ReadMe.txt
@@ -1,23 +1,0 @@
-Set path in the command prompt to the java bin.
-
-Extract the contents of the zip into a folder and set path to that folder
-
-Execute the following commands
-
-javac NeuronLayer1.java
-javac NeuronLAyer2.java
-javac NeuronLayer3.java
-javac Network.java
-
-Then execute 
-java Network
-
-Tha accuracy for 1 epoch and 100 epoch will be displayed.
-
-Note: Set path to where you have stored the dataset file and where you want the output in the 'Network.java' file at line numbers
-55
-57
-133
-135
-
-
