+++ src/NeuralNetwork.java
@@ -1,0 +1,223 @@
+package src;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.FileReader;
+import java.io.FileWriter;
+import java.util.ArrayList;
+
+/**
+ * Created by Matt on 2015-06-07.
+ */
+
+class NeuralNetwork {
+    private static InputLayer n1;
+    private static InputLayer n2;
+    private static HiddenLayer n3;
+    private static HiddenLayer n4;
+    private static OutputLayer n5;
+    private static InputLayer bias1;
+    private static HiddenLayer bias2;
+    private static final double learning_rate = 0.05;
+    static ArrayList<Double> errors = new ArrayList<Double>();
+
+    void initializeNeuralNetwork() {
+        n1 = new InputLayer();
+        n2 = new InputLayer();
+        n3 = new HiddenLayer();
+        n4 = new HiddenLayer();
+        n5 = new OutputLayer();
+        bias1 = new InputLayer();
+        bias2 = new HiddenLayer();
+        n1.setWeight1(0.33);
+        n1.setWeight2(0.42);
+        n2.setWeight1(0.5);
+        n2.setWeight2(-0.17);
+        n3.setWeight1(0.35);
+        n4.setWeight1(0.18);
+        bias1.setWeight1(0.29);
+        bias1.setWeight2(0.1);
+        bias2.setWeight1(-0.43);
+        bias1.setInput(1);
+        bias1.setOutput();
+        bias2.setInput(1);
+        bias2.setOutput();
+    } // END initializeNeuralNetwork()
+
+    public static void main(String[] args) {
+        NeuralNetwork neural = new NeuralNetwork();
+        neural.initializeNeuralNetwork();
+
+        double correct_for_1epoch = 0;
+        double correct_for_100epoch = 0;
+        int lines_count = 0;
+
+        //for 1 epoch
+        try {
+            FileReader fr = new FileReader(".\\src\\dataset.txt");
+            BufferedReader br = new BufferedReader(fr);
+            FileWriter fw = new FileWriter(".\\src\\output1.txt");
+            BufferedWriter bw = new BufferedWriter(fw);
+            String line;
+
+            while((line = br.readLine()) != null) {
+                lines_count++;
+                String [] lines = line.split(" ");
+                n1.setInput(Integer.parseInt(lines[0]));
+                n1.computeOutput();
+                n2.setInput(Integer.parseInt(lines[1]));
+                n2.computeOutput();
+
+                double layer1_output1 = n1.getOutput()*n1.getWeight1() + n2.getOutput()*n2.getWeight1() + bias1.getOutput()*bias1.getWeight1();
+                double layer1_output2 = n1.getOutput()*n1.getWeight2() + n2.getOutput()*n2.getWeight2() + bias1.getOutput()*bias1.getWeight2();
+
+                n3.setInput(layer1_output1);
+                n3.computeOutput();
+                n4.setInput(layer1_output2);
+                n4.computeOutput();
+
+                double output_layer2 = n3.getOutput()*n3.getWeight1() + n4.getOutput()*n4.getWeight1() + bias2.getOutput()*bias2.getWeight1();
+
+                n5.setInput(output_layer2);
+                n5.computeOutput();
+                int c = 0;
+
+                double final_output = n5.getOutput();
+
+                if(final_output >= 0.5) {
+                    c = 1;
+                    bw.write("1");
+                    bw.newLine();
+                } else {
+                    bw.write("0");
+                    bw.newLine();
+                }
+                double target = Integer.parseInt(lines[2]);
+
+                if(c == target) {
+                    correct_for_1epoch++;
+                }
+
+                double error = target - final_output;
+                double val_for_weight1_layer2 = learning_rate * error * n3.getOutput() * n5.getOutput() * (1 - n5.getOutput());
+                double val_for_weight2_layer2 = learning_rate * error * n4.getOutput() * n5.getOutput() * (1 - n5.getOutput());
+                double val_for_bias2 = learning_rate * error * bias2.getOutput() * n5.getOutput() * (1 - n5.getOutput());
+                bias2.setWeight1(bias2.getWeight1() + val_for_bias2);
+                n3.setWeight1(n3.getWeight1() + val_for_weight1_layer2);
+                n4.setWeight1(n4.getWeight1() + val_for_weight2_layer2);
+
+                double val_for_weight1_layer1 = learning_rate * error * n1.getOutput() * n3.getOutput() * (1 - n3.getOutput());
+                n1.setWeight1(n1.getWeight1() + val_for_weight1_layer1);
+                double val_for_weight2_layer1 = learning_rate * error * n1.getOutput() * n4.getOutput() * (1 - n4.getOutput());
+                n1.setWeight2(n1.getWeight2() + val_for_weight2_layer1);
+                double val_for_weight3_layer1 = learning_rate * error * n2.getOutput() * n3.getOutput() * (1 - n3.getOutput());
+                n2.setWeight1(n2.getWeight1() + val_for_weight3_layer1);
+                double val_for_weight4_layer1 = learning_rate * error * n2.getOutput() * n4.getOutput() * (1 - n4.getOutput());
+                n2.setWeight2(n2.getWeight2() + val_for_weight4_layer1);
+                double val_for_bias11 = learning_rate * error * bias1.getOutput() * n3.getOutput() * (1 - n3.getOutput());
+                bias1.setWeight1(bias1.getWeight1() + val_for_bias11);
+                double val_for_bias12 = learning_rate * error * bias1.getOutput() * n4.getOutput() * (1 - n4.getOutput());
+                bias1.setWeight2(bias1.getWeight1() + val_for_bias12);
+            }
+            bw.close();
+            br.close();
+            double acc = (correct_for_1epoch/lines_count) * 100;
+            System.out.println("Accuracy for 1 epoch = " + acc);
+            System.out.println("Final weights for node1  " + n1.getWeight1() + "  and " + n1.getWeight2());
+            System.out.println("Final weights for node2  " + n2.getWeight1() + "  and  " + n2.getWeight2());
+            System.out.println("Final weights for node3  " + n3.getWeight1());
+            System.out.println("Final weights for node4  " + n4.getWeight1());
+            System.out.println("Final weights for bias1  " + bias1.getWeight1() + "  and  " + bias1.getWeight2());
+            System.out.println("Final weights for bias2  " + bias2.getWeight1());
+        } catch(Exception e) {
+            e.printStackTrace();
+        }
+
+        //for 100 epoch
+        for(int i = 0; i < 100; i++) {
+            try {
+                FileReader fr = new FileReader(".\\src\\dataset.txt");
+                BufferedReader br = new BufferedReader(fr);
+                FileWriter fw = new FileWriter(".\\src\\output2.txt");
+                BufferedWriter bw = new BufferedWriter(fw);
+                String line;
+
+                while((line = br.readLine()) != null) {
+                    String [] lines = line.split(" ");
+                    n1.setInput(Integer.parseInt(lines[0]));
+                    n1.computeOutput();
+                    n2.setInput(Integer.parseInt(lines[1]));
+                    n2.computeOutput();
+
+                    double layer1_output1 = n1.getOutput()*n1.getWeight1() + n2.getOutput()*n2.getWeight1() + bias1.getOutput()*bias1.getWeight1();
+                    double layer1_output2 = n1.getOutput()*n1.getWeight2() + n2.getOutput()*n2.getWeight2() + bias1.getOutput()*bias1.getWeight2();
+
+                    n3.setInput(layer1_output1);
+                    n3.computeOutput();
+                    n4.setInput(layer1_output2);
+                    n4.computeOutput();
+
+                    double output_layer2 = n3.getOutput()*n3.getWeight1() + n4.getOutput()*n4.getWeight1() + bias2.getOutput()*bias2.getWeight1();
+
+                    n5.setInput(output_layer2);
+                    n5.computeOutput();
+                    int c = 0;
+
+                    double final_output = n5.getOutput();
+                    double target = Integer.parseInt(lines[2]);
+
+                    if(i == 99) {
+                        if(final_output >= 0.5) {
+                            c = 1;
+                            bw.write("1");
+                            bw.newLine();
+                        } else{
+                            bw.write("0");
+                            bw.newLine();
+                        }
+                        if(c == target) {
+                            correct_for_100epoch++;
+                        }
+                    }
+
+                    double error = target - final_output;
+                    double val_for_weight1_layer2 = learning_rate * error * n3.getOutput() * n5.getOutput() * (1 - n5.getOutput());
+                    double val_for_weight2_layer2 = learning_rate * error * n4.getOutput() * n5.getOutput() * (1 - n5.getOutput());
+                    double val_for_bias2 = learning_rate * error * bias2.getOutput() * n5.getOutput() * (1 - n5.getOutput());
+                    bias2.setWeight1(bias2.getWeight1() + val_for_bias2);
+                    n3.setWeight1(n3.getWeight1() + val_for_weight1_layer2);
+                    n4.setWeight1(n4.getWeight1() + val_for_weight2_layer2);
+
+                    double val_for_weight1_layer1 = learning_rate * error * n1.getOutput() * n3.getOutput() * (1 - n3.getOutput());
+                    n1.setWeight1(n1.getWeight1() + val_for_weight1_layer1);
+                    double val_for_weight2_layer1 = learning_rate * error * n1.getOutput() * n4.getOutput() * (1 - n4.getOutput());
+                    n1.setWeight2(n1.getWeight2() + val_for_weight2_layer1);
+                    double val_for_weight3_layer1 = learning_rate * error * n2.getOutput() * n3.getOutput() * (1 - n3.getOutput());
+                    n2.setWeight1(n2.getWeight1() + val_for_weight3_layer1);
+                    double val_for_weight4_layer1 = learning_rate * error * n2.getOutput() * n4.getOutput() * (1 - n4.getOutput());
+                    n2.setWeight2(n2.getWeight2() + val_for_weight4_layer1);
+                    double val_for_bias11 = learning_rate * error * bias1.getOutput() * n3.getOutput() * (1 - n3.getOutput());
+                    bias1.setWeight1(bias1.getWeight1() + val_for_bias11);
+                    double val_for_bias12 = learning_rate * error * bias1.getOutput() * n4.getOutput() * (1 - n4.getOutput());
+                    bias1.setWeight2(bias1.getWeight1() + val_for_bias12);
+                }
+                bw.close();
+                br.close();
+
+                if(i == 99) {
+                    double acc = (correct_for_100epoch/lines_count) * 100;
+                    System.out.println("Accuracy for 100 epoch = " + acc);
+
+                    System.out.println("Final weights for node1  " + n1.getWeight1() + "  and " + n1.getWeight2());
+                    System.out.println("Final weights for node2  " + n2.getWeight1() + "  and  " + n2.getWeight2());
+                    System.out.println("Final weights for node3  " + n3.getWeight1());
+                    System.out.println("Final weights for node4  " + n4.getWeight1());
+                    System.out.println("Final weights for bias1  " + bias1.getWeight1() + "  and  " + bias1.getWeight2());
+                    System.out.println("Final weights for bias2  " + bias2.getWeight1());
+                }
+            } catch(Exception e){
+                e.printStackTrace();
+            }
+        }
+    }
+} // END NeuralNeuralNetwork.java
