+++ src/output1.txt
@@ -1,0 +1,16 @@
+1
+1
+1
+1
+1
+1
+1
+1
+1
+0
+0
+0
+1
+1
+1
+1
